clarified used of foxglove:// links being passed into binary (#252)
### Description

Added context on how to pass a `foxglove://` link to the app binary when
opening. Also added mention of this in the CLI section of docs for
discoverability purposes - reasonable that is somewhere one might look
if wanting to open data sources via command line. Made clear this isn't
a function of the `foxglove` CLI tool.

---------

Co-authored-by: Jacob Bandes-Storch <jacob@foxglove.dev>